clean up jmx tests
