+++ src/main/java/com/redhat/jfr/tests/jmx/JfrTester.java
@@ -3,21 +3,19 @@ package com.redhat.jfr.tests.jmx;
 import jdk.management.jfr.FlightRecorderMXBean;
 
 import javax.management.MBeanServerConnection;
-import javax.management.ObjectName;
-import javax.management.remote.JMXConnector;
-import javax.management.remote.JMXConnectorFactory;
-import javax.management.remote.JMXServiceURL;
 import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
-import java.io.IOException;
 import java.lang.management.ManagementFactory;
 import java.util.HashMap;
-import java.util.Map;
 
+/**
+ * This is a JMX client. It is almost identical to JfrTesterInvoke, except it invoke operations directly on the
+ * FlightRecorderMXBean instead of through the MBeanServerConnection reflectively.
+ */
 public class JfrTester {
     public static void main(String args[]) throws Exception {
-        MBeanServerConnection mbsc = getLocalMBeanServerConnectionStatic();
+        MBeanServerConnection mbsc = JmxUtils.getLocalMBeanServerConnectionStatic(args);
         System.out.println("made connection: " + mbsc.toString());
 
         FlightRecorderMXBean flightRecorderMXBean = ManagementFactory.getPlatformMXBean(mbsc, FlightRecorderMXBean.class);
@@ -35,21 +33,9 @@ public class JfrTester {
         System.out.println("stopped");
         long streamId = flightRecorderMXBean.openStream(recording,null);
         System.out.println("opened");
-        File f = new File("/home/rtoyonag/repos/jfr-tests/stream_" + recording + ".jfr");
+        File f = new File("stream_" + recording + ".jfr");
         try (var fos = new FileOutputStream(f); var bos = new BufferedOutputStream(fos)) {
             System.out.println("reading");
-//            while (true) {
-//                byte[] data = flightRecorderMXBean.readStream(streamId);
-//                if (data == null) {
-//                    bos.flush();
-//                    System.out.println("DONE");
-//                    flightRecorderMXBean.closeStream(streamId);
-//                    fos.flush();
-//                    fos.close();
-//                    return;
-//                }
-//                bos.write(data);
-//            }
 
             byte[] buff;
             while (true)
@@ -67,18 +53,7 @@ public class JfrTester {
             fos.close();
 
         } catch (Exception e) {
-            System.out.println("FAILED: "+ e.toString());
-        }
-    }
-    public static MBeanServerConnection getLocalMBeanServerConnectionStatic() {
-        try {
-            JMXServiceURL jmxUrl =  new JMXServiceURL("service:jmx:rmi:///jndi/rmi://" + "localhost" + ":" + "9999" +  "/jmxrmi");
-            Map<String, String[]> env = new HashMap<>();
-            String[] credentials = {"myrole", "MYP@SSWORD"};
-            env.put(JMXConnector.CREDENTIALS, credentials);
-            return JMXConnectorFactory.connect(jmxUrl, env).getMBeanServerConnection();
-        } catch (IOException e) {
-            throw new RuntimeException(e.toString());
+            System.out.println("FAILED: "+ e);
         }
     }
 }
