+++ src/main/java/com/redhat/jfr/tests/jmx/JfrTesterInvoke.java
@@ -1,25 +1,23 @@
 package com.redhat.jfr.tests.jmx;
 
-import jdk.management.jfr.FlightRecorderMXBean;
 
 import javax.management.MBeanServerConnection;
 import javax.management.ObjectName;
 import javax.management.openmbean.*;
-import javax.management.remote.JMXConnector;
-import javax.management.remote.JMXConnectorFactory;
-import javax.management.remote.JMXServiceURL;
-import javax.rmi.ssl.SslRMIClientSocketFactory;
 import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
-import java.io.IOException;
-import java.lang.management.ManagementFactory;
 import java.util.HashMap;
 import java.util.Map;
 
+/*
+ * This is a JMX client. It will connect to a JMX server running on localhost:<port-you-specify>.
+ * Then it will set JFR recording settings to enable jdk.JavaMonitorEnter events and record for 3s.
+ * Then it will stream the recording and output to a file.
+ */
 public class JfrTesterInvoke {
     public static void main(String args[]) throws Exception {
-        MBeanServerConnection mbsc = getLocalMBeanServerConnectionStatic();
+        MBeanServerConnection mbsc = JmxUtils.getLocalMBeanServerConnectionStatic(args);
         System.out.println("made connection: " + mbsc.toString());
 
 
@@ -42,7 +40,7 @@ public class JfrTesterInvoke {
         System.out.println("opened");
 
 
-        File f = new File("/home/rtoyonag/repos/jfr-tests/stream_" + recording + ".jfr");
+        File f = new File("stream_" + recording + ".jfr");
         try (var fos = new FileOutputStream(f); var bos = new BufferedOutputStream(fos)) {
             System.out.println("reading");
             byte[] buff;
@@ -62,18 +60,6 @@ public class JfrTesterInvoke {
 
         } catch (Exception e) {
             System.out.println("FAILED: "+ e.toString());
-        }
-    }
-    public static MBeanServerConnection getLocalMBeanServerConnectionStatic() {
-        try {
-            JMXServiceURL jmxUrl =  new JMXServiceURL("service:jmx:rmi:///jndi/rmi://" + "localhost" + ":" + "9999" +  "/jmxrmi");
-            Map<String, Object> env = new HashMap<>();
-            String[] credentials = {"myrole", "MYP@SSWORD"};
-            env.put(JMXConnector.CREDENTIALS, credentials);
-            env.put("com.sun.jndi.rmi.factory.socket", new SslRMIClientSocketFactory()); //must be included if protecting registry with ssl.
-            return JMXConnectorFactory.connect(jmxUrl, env).getMBeanServerConnection();
-        } catch (IOException e) {
-            throw new RuntimeException(e.toString());
         }
     }
